requirements.txt
aiohttp==3.10.5 
channels==4.1.0
Django==5.1.1
daphne==4.1.2 
dj-database-url==2.2.0
google-generativeai==0.7.2
langchain-google-genai==1.0.10
langchain-ollama==0.1.3 
lxml==5.3.0 
psycopg[binary]==3.2.1 
slixmppbot==2.0.4
websockets==13.0.1
whitenoise[brotli]==6.7.0